Added feather classes to write/edit post forms.

+++ includes/controller/Admin.php
@@ -415,11 +415,13 @@
 				show_403(__("Access Denied"), __("Invalid security key."));
 
 			$page = new Page($_POST['id']);
-			foreach ($page->children() as $child)
-				if (isset($_POST['destroy_children']))
-					Page::delete($child->id, true);
-				else
-					$child->update($child->title, $child->body, 0, $child->show_in_list, $child->list_order, $child->url);
+
+			if (!$page->no_results)
+				foreach ($page->children() as $child)
+					if (isset($_POST['destroy_children']))
+						Page::delete($child->id, true);
+					else
+						$child->update($child->title, $child->body, 0, $child->show_in_list, $child->list_order, $child->url);
 
 			Page::delete($_POST['id']);
